fi xci

+++ crates/swc_ecma_lexer/src/common/lexer/mod.rs
@@ -2055,7 +2055,7 @@ pub trait Lexer<'a, TokenAndSpan>: Tokens<TokenAndSpan> + Sized {
     fn read_str_lit(&mut self) -> LexResult<Self::Token> {
         debug_assert!(self.cur() == Some(b'\'') || self.cur() == Some(b'"'));
         let start = self.cur_pos();
-        let quote = self.cur().unwrap() as u8;
+        let quote = self.cur().unwrap();
 
         self.bump(); // '"' or '\''
